refactor: component names
