+++ src/components/Layout.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import styled from '@emotion/styled'
-import NavigationHeader from './navigationHeader'
+import NavigationHeader from './NavigationHeader'
 
 const MainWrapper = styled.div`
   display: flex;
