+++ src/pages/resume.mdx
@@ -1,5 +1,5 @@
-import Avatar from '../components/avatar'
-import SEO from '../components/seo'
+import Avatar from '../components/Avatar'
+import SEO from '../components/SEO'
 import styled from '@emotion/react'
 import { css } from '@emotion/react'
